+++ NEWS.md
@@ -2,6 +2,23 @@
 title: "NEWS for the emmeans package"
 ---
 
+## emmeans 1.7.4.999xxxx
+  * Modified default behavior of many methods for class `emm_list`,
+    and added more complete documentation. Many users start with
+    objects like `emm <- emmeans(model, pairwise ~ treatment)` and then
+    expect to use something like `plot(emm)` when what they want is
+    `plot(emm[[1]])`. For situations where this is likely, we now
+    default `which` to `I_bet(1)`, which shows a message and sets `which` to 1.
+    Also, when `which` is of length 1, we return that one result rather than a
+    list.
+  * We also added the same object[[I_bet(1)]] default to functions like
+    `add_grouping()`, `emmip()`, and `emmeans)` itself when they are called with
+    an `emm_list` object. Basically, lots of users are not attuned to the
+    distinction between an `emmGrid` and an `emm_list`, and expect these functions
+    to work. So we need to just try to accommodate them as much as
+    possible.
+    
+
 ## emmeans 1.7.4
   * Added an argument `enhance.levels` to `contrast()` that allows
     better labeling of the levels being contrasted. For example, now